refactor: modularise cht session
